+++ examples/pstring/stage1.cpp
@@ -3,7 +3,7 @@
 #include <supportlib/framer.hpp>
 #include <supportlib/assert.hpp>
 
-#include <test_resource.hpp>
+#include <pmr/test_resource.hpp>
 
 void test(bool verbose)
 {
