--- include/CMakeLists.txt
@@ -1,24 +1,0 @@
-set(CMAKE_CXX_STANDARD 23)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-if (MSVC)
-    add_definitions (
-        # Disable Microsoft's Secure STL.
-        /D_ITERATOR_DEBUG_LEVEL=0
-        # Use multiple processes for compiling.
-        /MP
-    )
-
-add_definitions (
-        # "qualifier applied to function type has no meaning; ignored"
-        /wd4180
-        #  integral constant overflow
-        /wd4307
-        # "'function': was declared deprecated" (referring to STL functions)
-        /wd4996
-    )
-
-endif()
-
-target_include_directories(pmr_test_resource PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
-
